Add PixelData and cleanup of PngLoader implementation

+++ src/graphics/Framebuffer.cpp
@@ -47,9 +47,9 @@ Framebuffer::Framebuffer(const char *apDevPath)
     ioctl(mFramebufferFile, FBIOGET_FSCREENINFO, &mFixedInfo);
 
     // set Canvas specific variables
-    mWidth = static_cast<int>(mVariableInfo.xres);
-    mHeight = static_cast<int>(mVariableInfo.yres);
-    mBytesPerPixel = static_cast<int>(mVariableInfo.bits_per_pixel / 8);
+    mWidth = mVariableInfo.xres;
+    mHeight = mVariableInfo.yres;
+    mBytesPerPixel = mVariableInfo.bits_per_pixel / 8;
     // std::clog << "Framebuffer opened. Width=" << mWidth << " Height=" << mHeight << " BytesPerPixel=" << mBytesPerPixel << std::endl;
 
     // set yres_virtual for double buffering
